requirements.txt
Scrapy==2.5.0
requests
python-decouple